getting rid of warnings - slowly progressing

+++ BlackVision/LibBlackVision/Source/Engine/Models/Plugins/Channels/Geometry/VertexAttributesChannelDescriptor.cpp
@@ -34,7 +34,7 @@ VertexAttributesChannelDescriptor::~VertexAttributesChannelDescriptor   ()
 //
 unsigned int VertexAttributesChannelDescriptor::GetNumVertexChannels	() const
 {
-    return m_attributeChannelDescriptors.size();
+    return (unsigned int) m_attributeChannelDescriptors.size();
 }
 
 // ************************************
